larger commit
this handles a few core functions. Creating a new user, logging in a user, refreshing a JWT token. We created a super user in our code by adding an env variable with an email, if this email matches someone signing up the owner flag is flipped to true. the first owner can create other owners in the future